style: Añade menú y reorganización de código
Se ha implementado como idea inicial un menú en el cual se añadirán las diferentes opciones contempladas en el panel de control. Además, se ha reorganizado el código para mantener  buenas prácticas.
