+++ decide/decide/settings.py
@@ -121,7 +121,7 @@ DATABASES = {
         'ENGINE': 'djongo',
         'NAME': 'prueba',
         'CLIENT': {
-           'host': '127.0.0.1',
+            'host': '127.0.0.1',
         }
     }
 }
