+++ decide_panel/src/css/App.css
@@ -5,17 +5,3 @@ html {
 html body {
   background-color: #141d22 !important;
 }
-
-#principal {
-  margin-top: 300px;
-  margin-left: 25%;
-}
-
-.card.jsx-3327870985 {
-  background: linear-gradient(
-    4deg,
-    #095656 -63.59%,
-    #000000 -20.3%,
-    #176ccf 70.46%
-  ) !important;
-}
