+++ decide_panel/src/css/ConnectionTest.css
@@ -1,0 +1,13 @@
+#card-connection-test {
+  margin-top: 300px;
+  margin-left: 25%;
+}
+
+.card.jsx-3327870985 {
+  background: linear-gradient(
+    4deg,
+    #095656 -63.59%,
+    #000000 -20.3%,
+    #176ccf 70.46%
+  ) !important;
+}
