+++ decide_panel/src/css/HeaderBar.css
@@ -1,0 +1,7 @@
+.tabview-demo .tabview-custom span {
+  margin: 0 0.5rem;
+}
+
+.p-tabview .p-tabview-panels {
+  background: none !important;
+}
